back-end/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Werkzeug==2.3.3
email-validator==2.1.0
PyJWT==2.8.0
gunicorn==21.2.0
PyPDF2==3.0.1
Pillow==10.3.0
pytesseract==0.3.10
psycopg2-binary